+++ tests/testthat/test-xgb.R
@@ -9,25 +9,42 @@ test_that("xgb.Booster + linear solver + predict() works", {
   # Load data
   data(agaricus.train)
   data(agaricus.test)
-  bst <- xgboost(data = agaricus.train$data,
-                 label = agaricus.train$label,
-                 eta = 1,
-                 nthread = 2,
-                 nrounds = 2,
-                 eval_metric = "logloss",
-                 objective = "binary:logistic",
-                 verbose = 0)
+  if (utils::packageVersion("xgboost") > "2.0.0.0") {
+    bst <- xgboost(
+      x = agaricus.train$data,
+      y = agaricus.train$label,
+      learning_rate = 1,
+      nthread = 2,
+      nrounds = 2,
+      eval_metric = "logloss",
+      objective = "reg:squarederror"
+    )
+  } else {
+    bst <- xgboost(
+      data = agaricus.train$data,
+      label = agaricus.train$label,
+      eta = 1,
+      nthread = 2,
+      nrounds = 2,
+      eval_metric = "logloss",
+      objective = "binary:logistic",
+      verbose = 0
+    )
+  }
   x <- axe_call(bst)
-  expect_equal(x$call, rlang::expr(dummy_call()))
+  if (utils::packageVersion("xgboost") > "2.0.0.0") {
+    extracted_call <- attr(x, "call")
+  } else {
+    extracted_call <- x$call
+  }
+  expect_equal(extracted_call, rlang::expr(dummy_call()))
   x <- axe_env(bst)
-  expect_lt(lobstr::obj_size(x), lobstr::obj_size(bst))
+  expect_lte(lobstr::obj_size(x), lobstr::obj_size(bst))
+  expect_lte(lobstr::obj_size(attributes(x)), lobstr::obj_size(attributes(bst)))
   x <- butcher(bst)
-  expect_equal(xgb.importance(model = x),
-               xgb.importance(model = bst))
-  expect_equal(predict(x, agaricus.test$data),
-               predict(bst, agaricus.test$data))
-  expect_equal(xgb.dump(x, with_stats = TRUE),
-               xgb.dump(bst, with_stats = TRUE))
+  expect_equal(xgb.importance(model = x), xgb.importance(model = bst))
+  expect_equal(predict(x, agaricus.test$data), predict(bst, agaricus.test$data))
+  expect_equal(xgb.dump(x, with_stats = TRUE), xgb.dump(bst, with_stats = TRUE))
 })
 
 test_that("xgb.Booster + tree-learning algo + predict() works", {
@@ -37,24 +54,45 @@ test_that("xgb.Booster + tree-learning algo + predict() works", {
   # Load data
   data(agaricus.train)
   data(agaricus.test)
-  dtrain <- xgb.DMatrix(data = agaricus.train$data,
-                        label = agaricus.train$label)
-  bst <- xgb.train(data = dtrain,
-                   booster = "gblinear",
-                   nthread = 2,
-                   nrounds = 2,
-                   eval_metric = "logloss",
-                   objective = "binary:logistic",
-                   print_every_n = 10000L)
+  dtrain <- xgb.DMatrix(
+    data = agaricus.train$data,
+    label = agaricus.train$label
+  )
+  if (utils::packageVersion("xgboost") > "2.0.0.0") {
+    bst <- xgb.train(
+      params = list(
+        booster = "gblinear",
+        nthread = 2,
+        eval_metric = "logloss",
+        objective = "binary:logistic",
+        print_every_n = 10000L
+      ),
+      nrounds = 2,
+      data = dtrain
+    )
+  } else {
+    bst <- xgb.train(
+      data = dtrain,
+      booster = "gblinear",
+      nthread = 2,
+      nrounds = 2,
+      eval_metric = "logloss",
+      objective = "binary:logistic",
+      print_every_n = 10000L
+    )
+  }
   x <- axe_call(bst)
-  expect_equal(x$call, rlang::expr(dummy_call()))
+  if (utils::packageVersion("xgboost") > "2.0.0.0") {
+    extracted_call <- attr(x, "call")
+  } else {
+    extracted_call <- x$call
+  }
+  expect_equal(extracted_call, rlang::expr(dummy_call()))
   x <- axe_env(bst)
-  expect_lt(lobstr::obj_size(x), lobstr::obj_size(bst))
+  expect_lte(lobstr::obj_size(x), lobstr::obj_size(bst))
+  expect_lte(lobstr::obj_size(attributes(x)), lobstr::obj_size(attributes(bst)))
   x <- butcher(bst)
-  expect_equal(xgb.importance(model = x),
-               xgb.importance(model = bst))
-  expect_equal(predict(x, agaricus.test$data),
-               predict(bst, agaricus.test$data))
-  expect_equal(xgb.dump(x, with_stats = TRUE),
-               xgb.dump(bst, with_stats = TRUE))
+  expect_equal(xgb.importance(model = x), xgb.importance(model = bst))
+  expect_equal(predict(x, agaricus.test$data), predict(bst, agaricus.test$data))
+  expect_equal(xgb.dump(x, with_stats = TRUE), xgb.dump(bst, with_stats = TRUE))
 })