+++ tests/testthat/test-xrf.R
@@ -9,7 +9,15 @@ test_that("xrf + axe_call() works", {
       family = 'gaussian'
     )
   x <- axe_call(res)
-  expect_equal(x$xgb$call, rlang::expr(dummy_call()))
+
+  # due to new xgboost version
+  # https://github.com/tidymodels/butcher/issues/294
+  if (is.null(x$xgb$call)) {
+    expect_equal(attr(x$xgb, "call"), rlang::expr(dummy_call()))
+  } else {
+    expect_equal(x$xgb$call, rlang::expr(dummy_call()))
+  }
+
   expect_equal(x$glm$model$glmnet.fit$call, rlang::expr(dummy_call()))
   expect_equal(x$glm$model$call, rlang::expr(dummy_call()))
 })
@@ -24,9 +32,16 @@ test_that("xrf + axe_env() works", {
     )
   x <- axe_env(res)
   expect_equal(attr(x$base_formula, ".Environment"), rlang::base_env())
-  expect_equal(attr(x$rule_augmented_formula, ".Environment"), rlang::base_env())
+  expect_equal(
+    attr(x$rule_augmented_formula, ".Environment"),
+    rlang::base_env()
+  )
   expect_equal(attr(x$glm$formula, ".Environment"), rlang::base_env())
-  expect_equal(environment(x$xgb$callbacks[[1]]), rlang::base_env())
+  # due to new xgboost version
+  # https://github.com/tidymodels/butcher/issues/294
+  if (!is.null(x$xgb$callbacks)) {
+    expect_equal(environment(x$xgb$callbacks[[1]]), rlang::base_env())
+  }
 })
 
 test_that("xrf + butcher() works", {
@@ -38,13 +53,27 @@ test_that("xrf + butcher() works", {
       family = 'gaussian'
     )
   x <- butcher(res)
-  expect_equal(x$xgb$call, rlang::expr(dummy_call()))
+  # due to new xgboost version
+  # https://github.com/tidymodels/butcher/issues/294
+  if (is.null(x$xgb$call)) {
+    expect_equal(attr(x$xgb, "call"), rlang::expr(dummy_call()))
+  } else {
+    expect_equal(x$xgb$call, rlang::expr(dummy_call()))
+  }
+
   expect_equal(x$glm$model$glmnet.fit$call, rlang::expr(dummy_call()))
   expect_equal(x$glm$model$call, rlang::expr(dummy_call()))
   expect_equal(attr(x$base_formula, ".Environment"), rlang::base_env())
-  expect_equal(attr(x$rule_augmented_formula, ".Environment"), rlang::base_env())
+  expect_equal(
+    attr(x$rule_augmented_formula, ".Environment"),
+    rlang::base_env()
+  )
   expect_equal(attr(x$glm$formula, ".Environment"), rlang::base_env())
-  expect_equal(environment(x$xgb$callbacks[[1]]), rlang::base_env())
+  # due to new xgboost version
+  # https://github.com/tidymodels/butcher/issues/294
+  if (!is.null(x$xgb$callbacks)) {
+    expect_equal(environment(x$xgb$callbacks[[1]]), rlang::base_env())
+  }
 })
 
 test_that("xrf + predict() works", {