+++ NEWS.md
@@ -4,6 +4,11 @@
 
 * Added methods for `rset`, `rsplit`, `tune_results`, and `workflow_set` classes (#292).
 
+* Make to work with new versions of xgboost models (#294).
+
+* Added butcher methods for `tabnet()` (@cregouby #226).
+
+
 # butcher 0.3.6
 
 * Updated methods for `kknn::kknn()` to no longer remove the call (#288).