+++ NEWS.md
@@ -1,4 +1,6 @@
-# butcher (development version)
+# butcher 0.4.0
+
+* Max Kuhn is now the maintainer.
 
 * Added methods for `MASS::polr` (@pbulsink, #289).